Enable py3 for fedora
wsme has py3 support, they are running py3 jobs upstream,
Let's reenable py3 for fedora.
Also disabled tests with python3, as upstream has upperbound
for sqlalchemy set to 0.7.99, while we have > 1.2.5.
In centos we are not hitting this currently because tests are not
running(because setuptools is 22.0.5), after updating it
we will hit in centos as well.

Change-Id: I2a111be3cd405e3577c058018e525701a6e1426d